test(user-service): UserServiceImpl - tests for deleteById method - tests for getById method - add existsById method in UserRepository - change deleteById method

+++ user-service/src/main/java/io/github/kstnnn/user/service/service/impl/UserServiceImpl.java
@@ -35,6 +35,11 @@ public class UserServiceImpl implements UserService {
 
   @Override
   public void deleteById(UUID id) {
+    log.info("Check if user exists");
+    var isExists = userRepository.existsById(id);
+    if (!isExists) {
+      throw new UserNotFoundException(id);
+    }
     log.info("Delete user with id {} attempt", id);
     userRepository.deleteById(id);
   }
